ls -lに必要な機能をモジュール化
メインのクラス内の記述を減らします。
