+++ 05.ls/ls.rb
@@ -3,10 +3,11 @@
 require 'optparse'
 require 'pathname'
 require 'io/console/size'
-require 'etc'
-require 'time'
+require_relative './lib/ls_l'
 
 class Ls
+  include LsL
+
   @options = {}
   @path_names = []
   @column_count = 3
@@ -58,39 +59,6 @@ class Ls
     def clear_path_names
       @path_names.clear
     end
-
-    def convert_file_type(file_type)
-      file_types = {
-        'file' => '-',
-        'directory' => 'd',
-        'characterSpecial' => 'c',
-        'blockSpecial' => 'b',
-        'fifo' => 'p',
-        'link' => 'l'
-      }
-
-      file_types[file_type]
-    end
-
-    def convert_permission(modes)
-      outputs = ''
-      permissions = {
-        7 => 'rwx',
-        6 => 'rw-',
-        5 => 'r-x',
-        4 => 'r--',
-        3 => '-wx',
-        2 => '-w-',
-        1 => '--x',
-        0 => '---'
-      }
-
-      modes.each do |mode|
-        outputs += permissions[mode]
-      end
-
-      outputs
-    end
   end
 
   attr_accessor :files, :list_height, :file_name_width, :total_size, :symbolic_link_width, :size_width
@@ -119,21 +87,6 @@ class Ls
     Ls.options[:r] ? files.reverse : files
   end
 
-  def load_file_properties(file_stat, file)
-    file_properties = {}
-
-    file_properties[:file_type] = Ls.convert_file_type(file_stat.ftype)
-    file_properties[:permission] = Ls.convert_permission(file_stat.mode.digits(8)[0..2].reverse)
-    file_properties[:symbolic_link] = file_stat.nlink
-    file_properties[:owner_name] = Etc.getpwuid(file_stat.uid).name
-    file_properties[:group_name] = Etc.getgrgid(file_stat.gid).name
-    file_properties[:size] = file_stat.size
-    file_properties[:mtime] = file_stat.mtime.strftime('%_2m %_d %H:%M')
-    file_properties[:name] = file
-    file_properties[:read_link] = File.readlink("#{@path_name}/#{file_name}") if file_stat.ftype == 'link'
-    file_properties
-  end
-
   def configure_list_height
     @list_height = (@files.size.to_f / Ls.column_count).ceil
   end
@@ -144,20 +97,7 @@ class Ls
 
   def print_files
     if Ls.options[:l]
-      outputs = []
-      @files.each do |file|
-        file_stat = File.lstat("#{@path_name}/#{file}")
-        @total_size += file_stat.blocks
-        outputs << load_file_properties(file_stat, file)
-        @symbolic_link_width = [outputs.last[:symbolic_link].digits.size, @symbolic_link_width].max
-        @size_width = [outputs.last[:size].digits.size, @size_width].max
-      end
-
-      puts "total #{@total_size}"
-
-      outputs.each do |output|
-        print_details(output)
-      end
+      ls_l
     else
       @list_height.times do |row|
         0...Ls.column_count.times do |column|
@@ -167,24 +107,6 @@ class Ls
         puts
       end
     end
-  end
-
-  def print_details(file_properties)
-    print file_properties[:file_type]
-    print file_properties[:permission]
-    print '  '
-    print format("%#{@symbolic_link_width}d", file_properties[:symbolic_link])
-    print ' '
-    print file_properties[:owner_name]
-    print '  '
-    print file_properties[:group_name]
-    print '  '
-    print format("%#{@size_width}d", file_properties[:size])
-    print ' '
-    print file_properties[:mtime]
-    print ' '
-    print file_properties[:name]
-    puts file_properties[:read_link] ? " -> #{file_properties[:read_link]}" : ''
   end
 end
